+++ lib/oci.rb
@@ -28,6 +28,7 @@ require 'oci/identity/identity'
 require 'oci/load_balancer/load_balancer'
 require 'oci/object_storage/object_storage'
 require 'oci/object_storage/transfer/transfer'
+require 'oci/resource_search/resource_search'
 
 # Top level module for the Oracle Cloud Infrastructure SDK
 module OCI