+++ lib/oci/resource_search/util.rb
@@ -1,5 +1,2 @@
 # Copyright (c) 2016, 2018, Oracle and/or its affiliates. All rights reserved.
 
-module OCI
-  VERSION = '2.3.3'.freeze
-end